Split client OAuth into its own library
