+++ cmake/coturnConfig.cmake.in
@@ -5,7 +5,7 @@
 #
 #    find_package(coturn)
 #    target_link_libraries(MY_TARGET_NAME ${coturn_LIBRARIES})
-#  
+#
 #  This file will define the following variables:
 #    coturn_FOUND: True if find coturn, other false
 #    coturn_LIBRARIES:  The list of all imported targets for coturn components
@@ -18,6 +18,7 @@ if (NOT coturn_FIND_COMPONENTS)
     set(coturn_FIND_COMPONENTS
         turncommon
         turnclient
+        turnclientoauth
         turn_server
 	)
 endif()
@@ -42,7 +43,7 @@ foreach(module ${coturn_FIND_COMPONENTS})
     if(${module}_FOUND)
         list(APPEND coturn_LIBRARIES coturn::${module})
     endif()
-    list(APPEND required "${module}_FOUND")  
+    list(APPEND required "${module}_FOUND")
 endforeach()
 
 # Run checks via find_package_handle_standard_args
