+++ src/CMakeLists.txt
@@ -3,6 +3,6 @@
 project(coturn)
 
 add_subdirectory(client)
+add_subdirectory(clientoauth)
 add_subdirectory(server)
 add_subdirectory(apps)
-
