+++ src/apps/oauth/oauth.c
@@ -29,8 +29,8 @@
  */
 
 #include "ns_turn_defs.h"     // for STRCPY, turn_time_t, uint8_t, uint32_t
-#include "ns_turn_msg.h"      // for convert_oauth_key_data, decode_oauth_t...
 #include "ns_turn_msg_defs.h" // for oauth_token, oauth_encrypted_block
+#include "ns_turn_oauth.h"    // for convert_oauth_key_data, decode_oauth_t...
 #include "ns_turn_utils.h"
 
 #include "apputils.h"
