+++ src/apps/uclient/mainuclient.c
@@ -29,6 +29,7 @@
  */
 
 #include "apputils.h"
+#include "ns_turn_oauth.h"
 #include "ns_turn_utils.h"
 #include "session.h"
 #include "uclient.h"
